apps/risk-service/requirements.txt
fastapi
uvicorn[standard]
redis
psycopg2-binary
pydantic
httpx
boto3
python-jose[cryptography]
python-multipart
sqlalchemy
numpy
scikit-learn
pandas
python-dotenv
